requirements.txt
-e .
aws-cdk.aws-lambda
aws-cdk.aws-sam
aws_lambda_powertools
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions
aws-cdk.aws-events-targets